add: 各種ネットワークのmintページで固有のメタデータが表示されるようにした
